corrected divisor value and converted divider method to return float instead of int

+++ NumberProblems-Objc/NumberProblems-Objc/AppDelegate.h
@@ -17,7 +17,7 @@
 
 - (void)iterateCount:(int)cnt;
 - (int)multiplied:(int)arg;
-- (int)divided:(int)arg;
+- (float)divided:(int)arg;
 
 @end
